% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheDirectory.R
\name{cacheDirectory}
\alias{cacheDirectory}
\title{Gesel cache directory}
\usage{
cacheDirectory(cache = NULL)
}
\arguments{
\item{cache}{String specifying the path to a cache directory.}
}
\value{
If \code{cache=NULL}, the path to the current cache directory is returned.

If \code{cache} is provided, it is set as the path to the cache directory, and the previous location is invisibly returned.
}
\description{
Get or set the location of the Gesel cache directory,
used as the default in functions like \code{\link{downloadGeneFile}} and \code{\link{downloadDatabaseFile}}.
}
\details{
The cache directory defaults to a location in the user cache directory, as chosen by the \pkg{rappdirs} package.
Users can modify this choice by setting the \code{GESEL_CACHE_DIRECTORY} environment variable before the first call to this function.
}
\examples{
cacheDirectory()

old <- cacheDirectory("/tmp/foo/bar")
cacheDirectory()

cacheDirectory(old)
cacheDirectory()

}
\author{
Aaron Lun
}
