% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidateRanges.R
\name{consolidateBlockSize}
\alias{consolidateBlockSize}
\title{Block size for consolidation}
\usage{
consolidateBlockSize(block.size = NULL)
}
\arguments{
\item{block.size}{Integer specifying the block size in bytes.
Larger sizes reduce the number of requests at the cost of increasing the size of each request.}
}
\value{
If \code{block.size=NULL}, the current block size is returned.

If \code{block.size} is provided, it is used to set the block size, and the previous value is returned invisibly.
}
\description{
Get or set the block size for consolidating HTTP range requests.
}
\details{
Each file is split up into blocks of size approximately equal to \code{consolidateBlockSize()}.
When performing a range request, all ranges in the same block will be retrieved. 
This consolidates near-adjacent ranges into a single request, reducing the number of requests at the cost of increasing the size of each request.

All ranges associated with a block will be cached in memory, even those that were not directly requested.
Subsequent function calls can then quickly retrieve ranges from this cache instead of making a new HTTP request. 
Downloading and caching the entire block also ensures that the same bytes will never be requested from the server twice in the same session.

Setting the block size to a value of 1 indicates that no consolidation is to be performed, i.e., each range is requested as-is.
}
\examples{
consolidateBlockSize()
old <- consolidateBlockSize(500)
consolidateBlockSize()
consolidateBlockSize(old)

}
\author{
Aaron Lun
}
