% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidateRanges.R
\name{consolidateMaxGap}
\alias{consolidateMaxGap}
\title{Maximum gap for consolidation}
\usage{
consolidateMaxGap(max.gap = NULL)
}
\arguments{
\item{max.gap}{Integer specifying the maximum size of the gap, in bytes.}
}
\value{
If \code{max.gap=NULL}, the current maximum gap is returned.

If \code{max.gap} is provided, it is used to set the maximum gap, and the previous maximum is returned invisibly.
}
\description{
Get or set the maximum gap between ranges of interest.
This determines how aggressively \code{\link{consolidateRanges}} will merge near-adjacent ranges together.
}
\examples{
consolidateMaxGap()
old <- consolidateMaxGap(500)
consolidateMaxGap()
consolidateMaxGap(old)

}
\author{
Aaron Lun
}
