% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidateRanges.R
\name{consolidateMaxUnused}
\alias{consolidateMaxUnused}
\title{Maximum proportion of unused bytes for consolidation}
\usage{
consolidateMaxUnused(max.unused = NULL)
}
\arguments{
\item{max.unused}{Number between 0 and 1 inclusive, specifying the maximum percentage of unused bytes.}
}
\value{
If \code{max.unused=NULL}, the current maximum proportion is returned.

If \code{max.unused} is provided, it is used to set the maximum proportion, and the previous maximum is returned invisibly.
}
\description{
Get or set the maximum proportion of unused bytes during consolidation.
This determines how aggressively \code{\link{consolidateRanges}} will merge near-adjacent ranges together.
}
\examples{
consolidateMaxUnused()
old <- consolidateMaxUnused(0.5)
consolidateMaxUnused()
consolidateMaxUnused(old)

}
\author{
Aaron Lun
}
