% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidateRanges.R
\name{consolidateRanges}
\alias{consolidateRanges}
\title{Consolidate near-adjacent ranges}
\usage{
consolidateRanges(boundaries, needed, max.gap = consolidateMaxGap())
}
\arguments{
\item{boundaries}{Integer vector containing the zero-indexed boundaries of the byte ranges.
This should have length equal to the number of ranges plus 1, where the \code{i}-th range is defined as \code{[boundaries[i], boundaries[i+1])}.}

\item{needed}{Integer vector specifying the ranges of interest.
Each entry should be an index into \code{boundaries}.}

\item{max.gap}{Number between 0 and 1 inclusive, specifying the maximum gap between the ranges of interest.
Larger values will produce a smaller number of larger ranges.}
}
\value{
List containing:
\itemize{
\item \code{start}, an integer vector containing the starts of the consolidated ranges.
\item \code{end}, an integer vector of length equal to \code{start}.
This contains the ends of the consolidated ranges.
\item \code{requested}, an integer vector of the individual ranges that will be requested after consolidation.
This will be a superset of \code{needed}, where a range is listed here if and only if it is enclosed within one of the consolidated ranges.
}
}
\description{
Consolidate near-adjacent byte ranges in HTTP range requests.
This reduces the number of individual requests at the cost of having larger requests.
}
\examples{
boundaries <- c(0, 10, 30, 100, 200, 210) # five ranges
consolidateRanges(boundaries, c(2, 3, 4))
consolidateRanges(boundaries, c(1, 3, 5))
consolidateRanges(boundaries, c(1, 5))
consolidateRanges(boundaries, c(1, 5), max.gap=100)

}
\author{
Aaron Lun
}
