% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDatabaseFile.R
\name{downloadDatabaseFile}
\alias{downloadDatabaseFile}
\alias{databaseUrl}
\title{Download Gesel database files}
\usage{
downloadDatabaseFile(
  name,
  url = databaseUrl(),
  cache = NULL,
  overwrite = FALSE
)

databaseUrl(url = NULL)
}
\arguments{
\item{name}{String containing the name of the file.
This usually has the species identifier as a prefix.}

\item{url}{String containing the base URL to the Gesel database files.}

\item{cache}{String specifying the path to a cache directory.
If \code{NULL}, a cache location is automatically chosen.}

\item{overwrite}{Logical scalar indicating whether any cached file should be overwritten.}
}
\value{
\code{downloadDatabaseFile} returns a string containing a path to the downloaded file.

For \code{databaseUrl}, if \code{url=NULL}, the function returns a string containing the URL to the Gesel database.
If \code{url} is provided, it instead stores \code{url} as the URL to the database, and the previous value of \code{url} is invisibly returned.
}
\description{
Default function to download Gesel database files.
}
\details{
The database URL defaults to the GitHub releases at \url{https://github.com/LTLA/gesel-feedstock}.
This can be altered by setting the \code{GESEL_DATABASE_URL} environment variable prior to the first call to this function.
}
\examples{
# Download file.
downloadDatabaseFile("9606_collections.tsv.gz")

# Altering the default database URL.
databaseUrl()
old <- databaseUrl("https://foo")
databaseUrl()
databaseUrl(old)

}
\author{
Aaron Lun
}
