% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadIndexRange.R
\name{downloadIndexRange}
\alias{downloadIndexRange}
\title{Download a byte range from a Gesel index file}
\usage{
downloadIndexRange(name, range, url = indexUrl())
}
\arguments{
\item{name}{String containing the name of the file.}

\item{range}{Integer vector of length 2 containing the zero-indexed half-open byte range to extract from the file,
i.e., \code{c(a, b)} should extract bytes from \eqn{[a, b)}.}

\item{url}{String containing the base URL to the Gesel indices.
For \code{indexUrl}, this can be \code{NULL} to set it back to the default.}
}
\value{
String containing the contents of the specified byte range.
}
\description{
Download a byte range from a file in the Gesel index.
}
\examples{
downloadIndexRange("9606_set2gene.tsv", c(0L, 100L))

}
\author{
Aaron Lun
}
