% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSetsForSomeGenes.R
\name{effectiveNumberOfGenes}
\alias{effectiveNumberOfGenes}
\title{Effective number of genes}
\usage{
effectiveNumberOfGenes(
  species,
  fetch = downloadDatabaseFile,
  fetch.args = list()
)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{fetch}{Function that accepts the name of a Gesel database file and returns an absolute path to that file.}

\item{fetch.args}{Named list of arguments to pass to \code{fetch}.}
}
\value{
Integer scalar specifying the number of genes in Gesel that belong to at least one set.
}
\description{
Count the number of genes in the Gesel database that belong to at least one set.
}
\details{
The return value should be used as the total number of balls when performing a hypergeometric test for gene set enrichment
(see \code{\link{phyper}}), instead of \code{nrow(fetchAllGenes(species))}.
This ensures that uninteresting genes like pseudo-genes or predicted genes are ignored during the calculation.
Otherwise, unknown genes would inappropriately increase the number of balls and understate the enrichment p-values.
}
\author{
Aaron Lun
}
