% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSomeCollections.R
\name{fetchCollectionSizes}
\alias{fetchCollectionSizes}
\title{Size of collections}
\usage{
fetchCollectionSizes(species, fetch = NULL, fetch.args = list())
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{fetch}{Function that accepts the name of a Gesel database file and returns an absolute path to that file.}

\item{fetch.args}{Named list of arguments to pass to \code{fetch}.}
}
\value{
Integer vector containing the size of each collection (i.e., the number of gene sets).
}
\description{
Quickly get the sizes of all gene set collections in the Gesel database.
This is more efficient than \code{\link{fetchAllCollections}} when only the sizes are of interest.
}
\examples{
head(fetchCollectionSizes("9606"))

}
\author{
Aaron Lun
}
