% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchGenesForAllSets.R
\name{fetchGenesForAllSets}
\alias{fetchGenesForAllSets}
\title{Fetch genes for all sets}
\usage{
fetchGenesForAllSets(species, cache = NULL, overwrite = FALSE)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{cache}{String containign the path to a cache directory.
If \code{NULL}, a cache location is automatically chosen.}

\item{overwrite}{Logical scalar indicating whether any cached file should be overwritten.}
}
\value{
List of integer vectors.
Each vector corresponds to a gene set in the same order as \code{\link{fetchAllSets}}.
Each vector contains the identities of the genes in that set, where each integer is an index into \code{\link{fetchAllGenes}}.
}
\description{
Fetch the identities for genes in all sets in the \pkg{gesel} index.
}
\examples{
all.sets <- fetchGenesForAllSets("9606")
length(all.sets)

# Genes in the first set:
fetchAllGenes("9606")$symbol[all.sets[[1]]]

# Details about the first set:
fetchAllSets("9606")[1,]

}
\author{
Aaron Lun
}
