% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchGeneTypes.R
\name{fetchGeneTypes}
\alias{fetchGeneTypes}
\title{Fetch available gene identfier types}
\usage{
fetchGeneTypes(species, config = NULL)
}
\arguments{
\item{species}{String specifying the taxonomy ID of the species of interest.}

\item{config}{Configuration list, typically created by \code{\link{newConfig}}.
If \code{NULL}, the default configuration is used.}
}
\value{
Character vector of the available types.
}
\description{
Fetch the list of available gene identifier types for this species.
}
\details{
In older versions (0.1.0) of the Gesel gene annotation file specification, an explicit listing of types was not generated,
so this function will just assume that \code{"ensembl"}, \code{"entrez"} and \code{"symbol"} are available.
}
\examples{
fetchGeneTypes("9606")
fetchGeneTypes("10090")

}
\author{
Aaron Lun
}
