% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchGeneVersion.R
\name{fetchGeneVersion}
\alias{fetchGeneVersion}
\title{Fetch version of the gene annotation files}
\usage{
fetchGeneVersion(species, config = NULL)
}
\arguments{
\item{species}{String specifying the taxonomy ID of the species of interest.}

\item{config}{Configuration list, typically created by \code{\link{newConfig}}.
If \code{NULL}, the default configuration is used.}
}
\value{
String containing the version of the Gesel gene annotation file specification.
}
\description{
Get the version of the Gesel gene annotation file specification used by the species of interest.
}
\examples{
fetchGeneVersion("9606")

}
\author{
Aaron Lun
}
