% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSetsForAllGenes.R
\name{fetchSetsForAllGenes}
\alias{fetchSetsForAllGenes}
\title{Fetch sets for all genes}
\usage{
fetchSetsForAllGenes(species, config = NULL)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{config}{Configuration list, typically created by \code{\link{newConfig}}.
If \code{NULL}, the default configuration is used.}
}
\value{
List of integer vectors.
Each vector corresponds to a gene, corresponding to a row of the data frame returned by \code{\link{fetchAllGenes}}.
Each vector contains the identities of the sets that contain that gene, 
where each integer is a set index that refers to a row of the data frame returned by \code{\link{fetchAllSets}}.
}
\description{
Fetch the identities of the sets that contain each gene in the Gesel database.
}
\details{
If this function is called once, the returned list will be cached in memory and re-used in subsequent calls to this function.
The cached data will also be used to speed up calls to \code{\link{fetchSetsForSomeGenes}}.
}
\examples{
all.genes <- fetchSetsForAllGenes("9606")
length(all.genes)

# Sets containing the first gene:
head(fetchAllSets("9606")[all.genes[[1]],])

# Details about the first gene:
fetchAllGenes("9606")$symbol[1]

}
\author{
Aaron Lun
}
