% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSetsForGene.R
\name{fetchSetsForGene}
\alias{fetchSetsForGene}
\title{Fetch sets for a single gene}
\usage{
fetchSetsForGene(
  species,
  gene,
  fetch.file = NULL,
  fetch.file.args = list(),
  fetch.range = NULL,
  fetch.range.args = list(),
  use.preloaded = TRUE
)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{gene}{Integer identifying the gene of interest, as an index into the list of all genes for this species 
(i.e., one of the lists returned by \code{\link{fetchAllGenes}}).}

\item{fetch.file}{Function that accepts the name of the file in the Gesel index and returns an absolute path to the file.
If \code{NULL}, it defaults to \code{\link{downloadIndexFile}}.}

\item{fetch.file.args}{Named list of arguments to pass to \code{fetch.file}.}

\item{fetch.range}{Function that accepts at least two arguments - 
the name of the file in the Gesel index, and an integer vector of length 2 containing the zero-indexed half-open byte range to extract from the file
(see \code{\link{downloadIndexRange}} for details).
It should return a string containing the contents of the specified byte range.
If \code{NULL}, it defaults to \code{\link{downloadIndexRange}}.}

\item{fetch.range.args}{Named list of arguments to pass to \code{fetch.file}.}

\item{use.preloaded}{Logical scalar indicating whether to use the preloaded value from a previous call to \code{\link{fetchSetsForAllGenes}}.}
}
\value{
Integer vector containing the identities of the sets that contain \code{gene},
where each integer is an index into \code{\link{fetchAllSets}}.
}
\description{
Fetch all sets that contain a single gene in the Gesel index.
}
\examples{
first.gene <- fetchSetsForGene("9606", 1)

# Sets containing the first gene
head(fetchAllSets("9606")[first.gene,])

}
\author{
Aaron Lun
}
