% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSomeSets.R
\name{fetchSetSizes}
\alias{fetchSetSizes}
\title{Size of gene sets}
\usage{
fetchSetSizes(species, config = NULL)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{config}{Configuration list, typically created by \code{\link{newConfig}}.
If \code{NULL}, the default configuration is used.}
}
\value{
Integer vector containing the size of each set (i.e., the number of gene sets).
}
\description{
Quickly get the size of the sets in the Gesel database.
This is more efficient than \code{\link{fetchAllSets}} when only the sizes are of interest.
}
\examples{
head(fetchSetSizes("9606"))

}
\author{
Aaron Lun
}
