% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSingleSet.R
\name{fetchSingleSet}
\alias{fetchSingleSet}
\title{Fetch a single gene set}
\usage{
fetchSingleSet(
  species,
  set,
  fetch.file = NULL,
  fetch.file.args = list(),
  fetch.range = NULL,
  fetch.range.args = list(),
  use.preloaded = TRUE
)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{set}{Integer identifying the set of interest, as an index into the list of all sets for this species 
(i.e., the return value of \code{\link{fetchAllSets}}).}

\item{fetch.file}{Function that accepts the name of the file in the Gesel index and returns an absolute path to the file.
If \code{NULL}, it defaults to \code{\link{downloadIndexFile}}.}

\item{fetch.file.args}{Named list of arguments to pass to \code{fetch.file}.}

\item{fetch.range}{Function that accepts at least two arguments - 
the name of the file in the Gesel index, and an integer vector of length 2 containing the zero-indexed half-open byte range to extract from the file
(see \code{\link{downloadIndexRange}} for details).
It should return a string containing the contents of the specified byte range.
If \code{NULL}, it defaults to \code{\link{downloadIndexRange}}.}

\item{fetch.range.args}{Named list of arguments to pass to \code{fetch.file}.}

\item{use.preloaded}{Logical scalar indicating whether to use the preloaded value from a previous call to \code{\link{fetchAllSets}}.}
}
\value{
Named list containing:
\itemize{
\item \code{name}, string containing the name of the gene set.
\item \code{description}, string containing a description for the gene set.
\item \code{size}, integer scalar specifying the number of genes in this gene set.
\item \code{collection}, integer scalar containing the index of the collection containing this gene set.
This index refers to a row of the output of \code{\link{fetchAllCollections}}.
\item \code{number}, integer scalar containing the index of the gene set inside the specified collection.
}
}
\description{
Fetch the details of a single gene set from the Gesel index.
This is more efficient than calling \code{\link{fetchAllSets}} when only one set is of interest.
}
\examples{
fetchSingleSet("9606", 1)

}
\author{
Aaron Lun
}
