% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSomeCollections.R
\name{fetchSomeCollections}
\alias{fetchSomeCollections}
\title{Fetch some collections}
\usage{
fetchSomeCollections(species, collections, config = NULL)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{collections}{Integer vector containing collection indices.
Each entry refers to a row of the data frame returned by \code{\link{fetchAllCollections}}).}

\item{config}{Configuration list, typically created by \code{\link{newConfig}}.
If \code{NULL}, the default configuration is used.}
}
\value{
Data frame with the same columns as the return value of \code{\link{fetchAllCollections}},
where each row corresponds to an entry of \code{collections}.
}
\description{
Fetch the details of some gene set collections from the Gesel database.
This can be more efficient than \code{\link{fetchAllCollections}} when only a few collections are of interest.
}
\details{
Every time this function is called, information from the requested \code{collections} will be added to an in-memory cache.
Subsequent calls to this function will re-use as many of the cached collections as possible.

If \code{\link{fetchAllCollections}} is called, information from all collections are cached in memory and will be retrieved when \code{fetchSomeCollections} is called.
If \code{collections} is large, it may be beneficial to call \code{\link{fetchAllCollections}} first before calling this function.
}
\examples{
fetchSomeCollections("9606", 1)

}
\author{
Aaron Lun
}
