% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSomeSets.R
\name{fetchSomeSets}
\alias{fetchSomeSets}
\title{Fetch some sets}
\usage{
fetchSomeSets(species, sets, config = NULL)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{sets}{Integer vector of set indices, where each set index refers to a row in the data frame returned by \code{\link{fetchAllSets}}.}

\item{config}{Configuration list, typically created by \code{\link{newConfig}}.
If \code{NULL}, the default configuration is used.}
}
\value{
Data frame with the same columns as the return value of \code{\link{fetchAllSets}},
where each row corresponds to an entry of \code{sets}.
}
\description{
Fetch the details of some gene sets from the Gesel database.
This can be more efficient than calling \code{\link{fetchAllSets}} when only a few sets are of interest.
}
\examples{
fetchSomeSets("9606", 1)

}
\author{
Aaron Lun
}
