% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{flushMemoryCache}
\alias{flushMemoryCache}
\title{Flush the in-memory cache}
\usage{
flushMemoryCache()
}
\value{
The in-memory cache is cleared.
\code{NULL} is invisibly returned.
}
\description{
Flush the in-memory cache for \pkg{gesel} data structures in the current R session.
}
\details{
The \pkg{gesel} package makes caches the data structures in the current R session to avoid unnecessary requests to the filesystem and remote server.
On rare occasion, these cached data structures may be out of date when the Gesel database files change.
In such cases, the cache can be flushed to ensure that the various \pkg{gesel} functions operate on the latest version of the database.
}
\examples{
flushMemoryCache()

}
\author{
Aaron Lun
}
