% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGenesByIdentifier.R
\name{mapGenesByIdentifier}
\alias{mapGenesByIdentifier}
\title{Map gene identifiers to indices}
\usage{
mapGenesByIdentifier(species, type, ignore.case = FALSE, more.args = list())
}
\arguments{
\item{species}{String specifying the taxonomy ID of the species of interest.}

\item{type}{String specifying the type of identifier.
This is typically one of \code{"symbol"}, \code{"entrez"}, and \code{"ensembl"},}

\item{ignore.case}{Logical scalar indicating whether case should be ignored.}

\item{more.args}{Named list of further arguments to pass to \code{\link{fetchAllGenes}}.}
}
\value{
Named list of integer vectors.
Each name corresponds to an identifier of the specified \code{type},
and each vector contains the genes associated with that identifier (after ignoring case, if \code{ignore.case=TRUE}).
Vector entries should be interpreted as indices into any of the lists returned by \code{\link{fetchAllGenes}}.
}
\description{
Create a mapping of gene identifiers (Ensembl, symbol, etc.) to their indices.
}
\examples{
mapping <- mapGenesByIdentifier("9606", type="symbol")

# Taking it for a spin:
found <- mapping[["SNAP25"]]
fetchAllGenes("9606")$symbol[found]

}
\author{
Aaron Lun
}
