% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConfig.R
\name{newConfig}
\alias{newConfig}
\title{Create a new configuration object}
\usage{
newConfig(
  fetch.gene = NULL,
  fetch.gene.args = list(),
  fetch.file = NULL,
  fetch.file.args = list(),
  fetch.ranges = NULL,
  fetch.ranges.args = list(),
  consolidate.block.size = NULL
)
}
\arguments{
\item{fetch.gene}{Function that accepts the name of a Gesel gene description file and returns an absolute path to the file.
If \code{NULL}, it defaults to \code{\link{downloadGeneFile}}.}

\item{fetch.gene.args}{Named list of arguments to pass to \code{fetch.gene}.}

\item{fetch.file}{Function that accepts the name of a Gesel database file and returns an absolute path to the file.
If \code{NULL}, it defaults to \code{\link{downloadDatabaseFile}}.}

\item{fetch.file.args}{Named list of arguments to pass to \code{fetch.file}.}

\item{fetch.ranges}{Function that accepts three arguments - 
the name of a Gesel database file, an integer vector containing the starts of the byte ranges, and another vector containing the ends of the byte ranges
(see \code{\link{downloadDatabaseRanges}} for details).
It should return a list of raw vectors with the contents of the specified byte ranges.
If \code{NULL}, it defaults to \code{\link{downloadDatabaseRanges}}.}

\item{fetch.ranges.args}{Named list of arguments to pass to \code{fetch.ranges}.}

\item{consolidate.block.size}{Integer specifying the block size for consolidated requests.
If \code{NULL}, it defaults to \code{\link{consolidateBlockSize}}.}
}
\value{
A list containing Gesel configuration settings.
}
\description{
Create a new configuration object to specify how the Gesel database should be queried.
This can be used by applications to point to a different Gesel database from the default.
}
\details{
The configuration list returned by \code{newConfig} can be passed to each \pkg{gesel} function to alter its behavior in a consistent manner.
For example, we can override \code{fetch.file} to retrieve database files from a shared filesystem instead of performing a HTTP request.

The configuration list also contains a cache of data structures that can be populated by \pkg{gesel} functions.
This avoids unnecessary fetch requests upon repeated calls to the same function.
If the cache becomes stale or too large, it can be cleared by calling \code{\link{flushMemoryCache}}.

If no configuration list is supplied to \pkg{gesel} functions, the default configuration is used.
The default is created by calling \code{newConfig} without any arguments.
}
\examples{
config <- newConfig()
 
}
\author{
Aaron Lun
}
