% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareDatabaseFiles.R
\name{prepareDatabaseFiles}
\alias{prepareDatabaseFiles}
\title{Prepare the Gesel database}
\usage{
prepareDatabaseFiles(
  species,
  collections,
  set.info,
  set.membership,
  num.genes,
  path = ".",
  validate = TRUE
)
}
\arguments{
\item{species}{String specifying the species in the form of its NCBI taxonomy ID.}

\item{collections}{Data frame of information about each gene set collection, where each row corresponds to a collection.
This data frame should contain the \code{title}, \code{description}, \code{source} and \code{maintainer} columns as described in \code{?\link{fetchAllCollections}}.}

\item{set.info}{List of data frames of length equal to \code{nrow(collections)}.
Each data frame corresponds to a collection where each row corresponds to a gene set.
Each data frame should have the \code{name} and \code{description} columns as described in \code{?\link{fetchAllSets}}.}

\item{set.membership}{List of list of integer vectors.
Each inner list corresponds to a collection and each vector corresponds to a gene set in that collection.
Each vector contains the identities of its constituent genes, as indices into the data frame returned by \code{\link{fetchAllGenes}}.
All gene indices should be positive and no greater than \code{num.genes}.}

\item{num.genes}{Integer specifying the total number of genes available for this species.}

\item{path}{String containing the path to a directory in which to create the database files.}

\item{validate}{Boolean indicating whether to run \code{\link{validateDatabaseFiles}} on the newly created files.}
}
\value{
Several files are produced at \code{path} with the \code{<species>_} prefix.
These can be made available for download with \code{\link{downloadDatabaseFile}}.
}
\description{
Prepare Gesel database files from various pieces of gene set information.
}
\examples{
# Mocking up some information.
collections <- data.frame(
    title=c("FOO", "BAR"),
    description=c("I am a foo", "I am a bar"),
    maintainer=c("Aaron", "Aaron"),
    source=c("https://foo", "https://bar")
)

set.info <- list(
    data.frame(
        name=sprintf("FOO_\%i", seq_len(20)),
        description=sprintf("this is FOO \%i", seq_len(20))
    ),
    data.frame(
        name=sprintf("BAR_\%i", seq_len(50)),
        description=sprintf("this is BAR \%i", seq_len(50))
    )
)

# Mocking up the gene sets.
num.genes <- 10000
set.membership <- list(
    lapply(seq_len(nrow(set.info[[1]])), function(i) {
        sample(num.genes, sample(num.genes, 1))
    }),
    lapply(seq_len(nrow(set.info[[2]])), function(i) {
        sample(num.genes, sample(num.genes, 1))
    })
)

# Now making the database files.
output <- tempfile()
dir.create(output)
prepareDatabaseFiles(
    "9606",
    collections, 
    set.info, 
    set.membership,
    num.genes,
    output
)

# We can then read directly from them:
config <- newConfig(fetch.file=function(x) file.path(output, x))
head(fetchAllSets("9606", config))

}
\author{
Aaron Lun
}
