% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querySets.R
\name{querySets}
\alias{querySets}
\title{Query gene sets}
\usage{
querySets(
  species,
  genes = NULL,
  text = NULL,
  types = NULL,
  counts.only = TRUE,
  config = NULL
)
}
\arguments{
\item{species}{String specifying the taxonomy ID of the species of interest.}

\item{genes}{Character vector of gene identifiers of any supported type.
These are typically Ensembl/Entrez identifiers or gene symbols.
If not \code{NULL}, this function will search for gene sets that overlap any of the supplied genes.}

\item{text}{String containing one or more keywords to search on, see the \code{query=} argument in \code{\link{searchSetText}}.
If not \code{NULL}, this function will search for gene sets that contain the (tokenized) keywords in their names or descriptions.}

\item{types}{Character vector specifying the types of gene identifiers to consider for \code{genes}.
This is typically one or more of \code{"symbol"}, \code{"entrez"}, and \code{"ensembl"}.}

\item{counts.only}{Boolean indicating whether to return a list of the overlapping genes in each set.
Only used if \code{genes} is provided.}

\item{config}{Configuration list, typically created by \code{\link{newConfig}}.
If \code{NULL}, the default configuration is used.}
}
\value{
A data frame containing one row per set that matches the query conditions.
This contains the following columns:
\itemize{
\item \code{name}, string containing the name of the gene set.
\item \code{description}, string containing a description of the gene set.
\item \code{size}, integer specifying the number of genes in this gene set.
\item \code{collection}, the name of the collection that contains this gene set.
\item \code{set}, integer specifying the Gesel set index that can be used in other \pkg{gesel} functions, e.g., \code{\link{fetchGenesForSomeSets}}.
}
If \code{genes} is provided, the data frame will additionally contain:
\itemize{
\item \code{count}, an integer column containing the number of overlaps between the genes in the set and those in \code{genes}.
\item \code{genes}, a nested list where each entry is a character vector containing the genes in \code{genes} that are present in each set.
Only reported if \code{counts.only = FALSE}.
\item \code{pvalue} column, a numeric column containing the hypergeometric p-value for overrepresentation of \code{genes} in the set.
Rows will be sorted by this column if it is present.
}
}
\description{
Query gene sets based on overlaps with genes of interest or matches to keyboards in their names/descriptions.
}
\details{
This is a user-friendly wrapper for quick and convenient searching of the Gesel database.
Developers may prefer to use the lower-level \pkg{gesel} functions for more customization and flexibility.
}
\examples{
out <- querySets(
    species = "9606",
    genes = c("tead1", "tead2", "tead3", "tead4"),
    text = "transcription"
)
head(out)

out2 <- querySets(
    species = "9606",
    genes = c("SNAP25", "neurod4", "neurod6"),
    text = "neuro*",
    counts.only = FALSE
)
head(out2)

}
\seealso{
\code{\link{searchGenes}}, to convert gene identifiers into internal Gesel indices.

\code{\link{searchOverlappingSets}}, to find the sets that overlap the genes of interest.

\code{\link{searchSetText}}, to find sets based on keywords in their names/descriptions.

\code{\link{fetchSomeSets}}, to get the details for each set. 

\code{\link{fetchSomeCollections}}, to get the details for each collection.
}
\author{
Aaron Lun
}
