% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDatabaseRanges.R
\name{rangeConcurrency}
\alias{rangeConcurrency}
\title{Concurrency of range requests}
\usage{
rangeConcurrency(concurrency = NULL)
}
\arguments{
\item{concurrency}{Integer containing the maximum number of concurrent requests per minute.}
}
\value{
If \code{concurrency=NULL}, the maximum number of concurrent requests is returned.

If \code{concurrency} is provided, it is set to the maximum number of concurrent requests, and the previous maximum is returned.
}
\description{
Get or set the maximum number of concurrent HTTP range requests that can be performed per minute in \code{\link{downloadDatabaseRanges}}.
Setting this to a smaller number avoids excessive load on the server.
}
\examples{
rangeConcurrency()
old <- rangeConcurrency(5)
rangeConcurrency()
rangeConcurrency(old)
rangeConcurrency()

}
\seealso{
\code{\link[httr2]{req_throttle}}, for the logic behind the requests-per-minute limit.
}
\author{
Aaron Lun
}
