% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDatabaseRanges.R
\name{readDatabaseRanges}
\alias{readDatabaseRanges}
\title{Read byte ranges from a Gesel database file}
\usage{
readDatabaseRanges(dir, name, start, end)
}
\arguments{
\item{dir}{String containing the path to a directory containing all Gesel database files.}

\item{name}{String containing the name of a Gesel database file.
This usually has the species identifier as a prefix, e.g., \code{"9606_set2gene.tsv.gz"}.}

\item{start}{Integer vector containing the zero-indexed closed start of each byte range to extract from the file.
This may be of zero length.}

\item{end}{Integer vector containing the zero-indexed open end of each byte range to extract from the file.
This should have the same length as \code{start} such that the \code{i}-th range is defined as \code{[start[i], end[i])}.
All ranges supplied in a single call to this function should be non-overlapping.}
}
\value{
List of length equal to \code{length(start)}.
Each entry is a raw vector representing the contents of the corresponding byte range.
}
\description{
Read any number of byte ranges from a Gesel database file on the local filesystem.
}
\examples{
path <- downloadDatabaseFile("9606_set2gene.tsv")
readDatabaseRanges(dirname(path), basename(path), 0L, 100L)
readDatabaseRanges(dirname(path), basename(path), c(10, 100, 1000), c(20, 150, 1100))

}
\seealso{
\code{\link{downloadDatabaseRanges}}, for a remote counterpart to this function that requests the byte ranges from a server.
}
