% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameGenesInSets.R
\name{renameGenesInSets}
\alias{renameGenesInSets}
\title{Rename genes in their sets}
\usage{
renameGenesInSets(species, sets, type, config = NULL)
}
\arguments{
\item{species}{String specifying the taxonomy ID of the species of interest.}

\item{sets}{List of integer vectors, typically generated by \code{\link{fetchGenesForAllSets}} or \code{\link{fetchGenesForSomeSets}}.
Each vector represents a gene set and contains gene indices, i.e., row indices into the data frame returned by \code{\link{fetchAllGenes}}.}

\item{type}{String specifying the type of gene identifier to convert to.
This is typically one of \code{"symbol"}, \code{"ensembl"} or \code{"entrez"}.}

\item{config}{Configuration list, typically created by \code{\link{newConfig}}.
If \code{NULL}, the default configuration is used.}
}
\value{
A list of the same length as \code{sets}.
Each entry is now a character vector containing the specified identifiers for the genes in the corresponding set.
}
\description{
Convert genes in sets from row indices to their identifiers.
}
\examples{
example.sets <- fetchGenesForSomeSets("9606", 1:50)
str(example.sets)
renamed.sets <- renameGenesInSets("9606", example.sets, "symbol")
str(renamed.sets)

}
\author{
Aaron Lun
}
