% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSetText.R
\name{searchSetText}
\alias{searchSetText}
\title{Search set text}
\usage{
searchSetText(
  species,
  query,
  use.name = TRUE,
  use.description = TRUE,
  fetch.file = NULL,
  fetch.file.args = list(),
  fetch.range = NULL,
  fetch.range.args = list()
)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{query}{String containing one or more words to search on.
A set is only matched if it matches to all of the tokens in the query.
The \code{*} and \code{?} wildcards can be used to match to any or one character, respectively.}

\item{use.name}{Logical scalar indicating whether to search on the name of the set.}

\item{use.description}{Logical scalar indicating whether to search on the description of the set.}

\item{fetch.file}{Function that accepts the name of the file in the Gesel database and returns an absolute path to the file.
If \code{NULL}, it defaults to \code{\link{downloadIndexFile}}.}

\item{fetch.file.args}{Named list of arguments to pass to \code{fetch.file}.}

\item{fetch.range}{Function that accepts at least two arguments - 
the name of the file in the Gesel database, and an integer vector of length 2 containing the zero-indexed half-open byte range to extract from the file
(see \code{\link{downloadIndexRange}} for details).
It should return a string containing the contents of the specified byte range.
If \code{NULL}, it defaults to \code{\link{downloadIndexRange}}.}

\item{fetch.range.args}{Named list of arguments to pass to \code{fetch.file}.}
}
\value{
Integer vector of set indices for the matching gene sets.
Each set index refers to a row in the data frame returned by \code{\link{fetchAllSets}}.
}
\description{
Search for sets based on their names and descriptions.
}
\examples{
out <- searchSetText("9606", "cancer")
fetchSomeSets("9606", out[1])

out <- searchSetText("9606", "innate immun*")
fetchSomeSets("9606", out[1])

}
\author{
Aaron Lun
}
