% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGeneIdentifierMap.R
\name{createGeneIdentifierMap}
\alias{createGeneIdentifierMap}
\alias{mapGenesByName}
\title{Create mapping of gene identifiers}
\usage{
createGeneIdentifierMap(species, type, ignore.case = FALSE, config = NULL)
}
\arguments{
\item{species}{String specifying the taxonomy ID of the species of interest.}

\item{type}{String specifying the type of gene identifier.
This is typically one of \code{"symbol"}, \code{"entrez"}, and \code{"ensembl"},}

\item{ignore.case}{Boolean indicating whether case should be ignored.}

\item{config}{Configuration list, typically created by \code{\link{newConfig}}.
If \code{NULL}, the default configuration is used.}
}
\value{
Named list of integer vectors.
Each name is an identifier of the specified \code{type},
and each vector contains the identities of genes associated with that identifier (after ignoring case, if \code{ignore.case=TRUE}).
Specifically, each gene's identity is represented as a row index into the data frame returned by \code{\link{fetchAllGenes}}.
}
\description{
Create a mapping of gene identifiers (Ensembl, symbol, etc.) to their Gesel gene indices.
}
\examples{
mapping <- createGeneIdentifierMap("9606", type="symbol")

# Taking it for a spin:
found <- mapping[["SNAP25"]]
fetchAllGenes("9606")$symbol[found]

}
\seealso{
\code{\link{searchGenes}}, which uses the mapping when searching for genes.
}
\author{
Aaron Lun
}
