% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDatabaseRanges.R
\name{downloadDatabaseRanges}
\alias{downloadDatabaseRanges}
\alias{downloadMultipartRanges}
\title{Fetch byte ranges from a Gesel database file}
\usage{
downloadDatabaseRanges(
  name,
  start,
  end,
  url = databaseUrl(),
  multipart = FALSE
)
}
\arguments{
\item{name}{String containing the name of the file.
This usually has the species identifier as a prefix.}

\item{start}{Integer vector containing the zero-indexed closed start of each byte range to extract from the file.
This may be of zero length.}

\item{end}{Integer vector containing the zero-indexed open end of each byte range to extract from the file.
This should have the same length as \code{start} such that the \code{i}-th range is defined as \code{[start[i], end[i])}.
All ranges supplied in a single call to this function should be non-overlapping.}

\item{url}{String containing the base URL to the Gesel database files.
For \code{databaseUrl}, this can be \code{NULL} to set it back to the default.}

\item{multipart}{Logical scalar indicating whether the server at \code{url} supports multi-part range requests.}
}
\value{
Character vector of length equal to \code{length(start)}, containing the contents of the requested byte ranges.
}
\description{
Download any number of byte ranges from a Gesel database file.
}
\examples{
downloadDatabaseRanges("9606_set2gene.tsv", 0L, 100L)
downloadDatabaseRanges("9606_set2gene.tsv", c(10, 100, 1000), c(20, 150, 1100))

}
\author{
Aaron Lun
}
