% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGeneFile.R
\name{downloadGeneFile}
\alias{downloadGeneFile}
\alias{geneUrl}
\title{Fetch Gesel gene description files}
\usage{
downloadGeneFile(name, url = geneUrl(), cache = NULL, overwrite = FALSE)

geneUrl(url)
}
\arguments{
\item{name}{String containing the name of the file, typically of the form \code{<species>_<type>.tsv.gz}, e.g., \code{"9606_symbol.tsv.gz"}.}

\item{url}{String containing the base URL to the Gesel gene descriptions. 
For \code{geneUrl}, this can be \code{NULL} to set it back to the default.}

\item{cache}{String specifying the path to a cache directory.
If \code{NULL}, a cache location is automatically chosen.}

\item{overwrite}{Logical scalar indicating whether any cached file should be overwritten.}
}
\value{
\code{downloadGeneFile} returns a string containing a local path to the downloaded file.

For \code{geneUrl}, if \code{url} is missing, the function returns a string containing the URL to the Gesel gene descriptions.
If \code{url} is provided, it instead stores \code{url} as the URL to the indices, and the previous value of \code{url} is invisibly returned.
}
\description{
Default function to download Gesel gene description files.
}
\details{
The default gene URL is set to the GitHub releases at \url{https://github.com/LTLA/gesel-feedstock}.
This can be altered by setting the \code{GESEL_GENE_URL} environment variable.

On first use of a cached file in an R session, \code{downloadGeneFile} will automatically check for updates at \code{url}. 
If the file on the remote has been updated, the new copy will be downloaded to the cache.
}
\examples{
# Download file.
downloadGeneFile("9606_symbol.tsv.gz")

# Altering the default gene URL.
geneUrl()
old <- geneUrl("https://foo")
geneUrl()
geneUrl(old)

}
\author{
Aaron Lun
}
