% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchAllCollections.R
\name{fetchAllCollections}
\alias{fetchAllCollections}
\title{Fetch all gene set collections}
\usage{
fetchAllCollections(species, cache = NULL, overwrite = FALSE)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{cache}{String containing the path to a cache directory.
If \code{NULL}, a cache location is automatically chosen.}

\item{overwrite}{Logical scalar indicating whether any cached file should be overwritten.}
}
\value{
Data frame of gene set collection information.
Each row represents a collection and contains:
\itemize{
\item \code{title}, string containing the title for the collection.
\item \code{description}, string containing a description for the collection.
\item \code{maintainer}, string containing the identity of the collection's maintainer.
\item \code{source}, string containing the source of origin of the collection.
\item \code{size}, integer scalar specifying the number of gene sets in the collection.
}
}
\description{
Fetch information about all gene set collections in the \pkg{gesel} index.
}
\examples{
out <- fetchAllCollections("9606")
head(out)

}
\author{
Aaron Lun
}
