% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchAllCollections.R
\name{fetchAllCollections}
\alias{fetchAllCollections}
\title{Fetch all gene set collections}
\usage{
fetchAllCollections(species, config = NULL)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{config}{Configuration list, typically created by \code{\link{newConfig}}.
If \code{NULL}, the default configuration is used.}
}
\value{
Data frame of gene set collection information.
Each row represents a collection and contains:
\itemize{
\item \code{title}, string containing the title for the collection.
\item \code{description}, string containing a description for the collection.
\item \code{maintainer}, string containing the identity of the collection's maintainer.
\item \code{source}, string containing the source of origin of the collection.
\item \code{start}, integer containing the set index for the first gene set in this collection.
The set index refers to a row in the data frame returned by \code{\link{fetchAllSets}}.
\item \code{size}, integer scalar specifying the number of gene sets in the collection.
}
}
\description{
Fetch information about all gene set collections in the Gesel database.
}
\examples{
out <- fetchAllCollections("9606")
head(out)

}
\author{
Aaron Lun
}
