% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchAllGenes.R
\name{fetchAllGenes}
\alias{fetchAllGenes}
\title{Fetch all genes}
\usage{
fetchAllGenes(species, types = NULL, cache = NULL, overwrite = FALSE)
}
\arguments{
\item{species}{String specifying the taxonomy ID of the species of interest.}

\item{types}{Character vector specifying the gene ID types to returnc.
This is typically one or more of \code{"symbol"}, \code{"entrez"}, and \code{"ensembl"},
defaulting to all of them.}

\item{cache}{String specifying the path to a cache directory.
If \code{NULL}, a cache location is automatically chosen.}

\item{overwrite}{Logical scalar indicating whether any cached file should be overwritten.}
}
\value{
List of length equal to \code{types}.
Each element is another list of length equal to the number of genes,
containing character vectors with the names of those genes.
}
\description{
Fetch names for all genes.
}
\examples{
out <- fetchAllGenes("9606")
names(out)
head(out$symbol)

}
\author{
Aaron Lun
}
