% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchAllGenes.R
\name{fetchAllGenes}
\alias{fetchAllGenes}
\title{Fetch all genes}
\usage{
fetchAllGenes(
  species,
  types = NULL,
  fetch = NULL,
  fetch.args = list(),
  use.preloaded = TRUE
)
}
\arguments{
\item{species}{String specifying the taxonomy ID of the species of interest.}

\item{types}{Character vector specifying the gene ID types to returnc.
This is typically one or more of \code{"symbol"}, \code{"entrez"}, and \code{"ensembl"},
defaulting to all of them.}

\item{fetch}{Function that accepts the name of the file in the Gesel gene descriptions and returns an absolute path to the file.
If \code{NULL}, it defaults to \code{\link{downloadGeneFile}}.}

\item{fetch.args}{Named list of arguments to pass to \code{fetch}.}

\item{use.preloaded}{Logical scalar indicating whether to use the preloaded value from a previous call to this function.}
}
\value{
List of length equal to \code{types}.
Each element is another list of length equal to the number of genes,
containing character vectors with the names of those genes.
}
\description{
Fetch names for all genes.
}
\examples{
out <- fetchAllGenes("9606")
names(out)
head(out$symbol)

}
\author{
Aaron Lun
}
