% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchAllGenes.R
\name{fetchAllGenes}
\alias{fetchAllGenes}
\title{Fetch all genes}
\usage{
fetchAllGenes(
  species,
  types = NULL,
  fetch = downloadGeneFile,
  fetch.args = list(),
  use.preloaded = TRUE
)
}
\arguments{
\item{species}{String specifying the taxonomy ID of the species of interest.}

\item{types}{Character vector specifying the types of gene names to return.
This is typically one or more of \code{"symbol"}, \code{"entrez"}, and \code{"ensembl"},
defaulting to all of them.}

\item{fetch}{Function that accepts the name of the file in the Gesel gene descriptions and returns an absolute path to the file.
If \code{NULL}, it defaults to \code{\link{downloadGeneFile}}.}

\item{fetch.args}{Named list of arguments to pass to \code{fetch}.}

\item{use.preloaded}{Logical scalar indicating whether to use the preloaded value from a previous call to this function.}
}
\value{
Data frame where each row represents a gene.
Each column corresponds to one of the \code{types} and is a list of character vectors.
Each vector in the column contains the names of the specified type for each gene.
}
\description{
Fetch names for all genes.
}
\examples{
out <- fetchAllGenes("9606")
head(out)
head(out$symbol)

}
\author{
Aaron Lun
}
