% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchAllSets.R
\name{fetchAllSets}
\alias{fetchAllSets}
\title{Fetch all gene sets}
\usage{
fetchAllSets(species, cache = NULL, overwrite = FALSE)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{cache}{String containing the path to a cache directory.
If \code{NULL}, a cache location is automatically chosen.}

\item{overwrite}{Logical scalar indicating whether any cached file should be overwritten.}
}
\value{
Data frame of gene set information.
Each row represents a gene set and contains:
\itemize{
\item \code{name}, string containing the name of the gene set.
\item \code{description}, string containing a description for the gene set.
\item \code{size}, integer scalar specifying the number of genes in this gene set.
\item \code{collection}, integer scalar containing the index of the collection containing this gene set.
This index refers to a row of the output of \code{\link{fetchAllCollections}}.
\item \code{number}, integer scalar containing the index of the gene set inside the specified collection.
}
}
\description{
Fetch information about all gene sets in the \pkg{gesel} index.
}
\examples{
out <- fetchAllSets("9606")
head(out)

}
\author{
Aaron Lun
}
