% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchAllSets.R
\name{fetchAllSets}
\alias{fetchAllSets}
\title{Fetch all gene sets}
\usage{
fetchAllSets(species, config = NULL)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{config}{Configuration list, typically created by \code{\link{newConfig}}.
If \code{NULL}, the default configuration is used.}
}
\value{
Data frame of gene set information.
Each row represents a gene set and contains:
\itemize{
\item \code{name}, string containing the name of the gene set.
\item \code{description}, string containing a description for the gene set.
\item \code{size}, integer scalar specifying the number of genes in this gene set.
\item \code{collection}, integer scalar containing the collection index of the collection that contains this gene set.
The collection index refers to a row of the data frame returned by \code{\link{fetchAllCollections}}.
\item \code{number}, integer scalar containing the position of the gene set inside the specified collection.
The set index of the current gene set is defined by adding \code{number - 1} to the collection's \code{start}. 
}
}
\description{
Fetch information about all gene sets in the Gesel database.
}
\examples{
out <- fetchAllSets("9606")
head(out)

}
\author{
Aaron Lun
}
