% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSomeCollections.R
\name{fetchCollectionSizes}
\alias{fetchCollectionSizes}
\title{Size of collections}
\usage{
fetchCollectionSizes(species, config = NULL)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{config}{Configuration list, typically created by \code{\link{newConfig}}.
If \code{NULL}, the default configuration is used.}
}
\value{
Integer vector containing the size of each collection (i.e., the number of gene sets).
}
\description{
Quickly get the sizes of all gene set collections in the Gesel database.
This is more efficient than \code{\link{fetchAllCollections}} when only the sizes are of interest.
}
\examples{
head(fetchCollectionSizes("9606"))

}
\author{
Aaron Lun
}
