% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSomeCollections.R
\name{fetchCollectionSizes}
\alias{fetchCollectionSizes}
\title{Size of collections}
\usage{
fetchCollectionSizes(
  species,
  fetch.file = NULL,
  fetch.file.args = list(),
  use.preloaded = TRUE
)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{fetch.file}{Function that accepts the name of the file in the Gesel database and returns an absolute path to the file.
If \code{NULL}, it defaults to \code{\link{downloadIndexFile}}.}

\item{fetch.file.args}{Named list of arguments to pass to \code{fetch.file}.}

\item{use.preloaded}{Logical scalar indicating whether to use the preloaded value from a previous call to \code{\link{fetchAllCollections}}.}
}
\value{
Integer vector containing the size of each collection (i.e., the number of gene sets).
}
\description{
Quickly get the sizes of all gene set collections in the Gesel database.
This is more efficient than \code{\link{fetchAllCollections}} when only the sizes are of interest.
}
\examples{
head(fetchCollectionSizes("9606"))

}
\author{
Aaron Lun
}
