% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchGenesForAllSets.R
\name{fetchGenesForAllSets}
\alias{fetchGenesForAllSets}
\title{Fetch genes for all sets}
\usage{
fetchGenesForAllSets(
  species,
  fetch = NULL,
  fetch.args = list(),
  use.preloaded = TRUE
)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{fetch}{Function that accepts the name of the file in the Gesel index and returns an absolute path to the file.
If \code{NULL}, it defaults to \code{\link{downloadIndexFile}}.}

\item{fetch.args}{Named list of arguments to pass to \code{fetch}.}

\item{use.preloaded}{Logical scalar indicating whether to use the preloaded value from a previous call to this function.}
}
\value{
List of integer vectors.
Each vector corresponds to a gene set in the same order as \code{\link{fetchAllSets}}.
Each vector contains the identities of the genes in that set, where each integer is an index into \code{\link{fetchAllGenes}}.
}
\description{
Fetch the identities for genes in all sets in the \pkg{gesel} index.
}
\examples{
all.sets <- fetchGenesForAllSets("9606")
length(all.sets)

# Genes in the first set:
fetchAllGenes("9606")$symbol[all.sets[[1]]]

# Details about the first set:
fetchAllSets("9606")[1,]

}
\author{
Aaron Lun
}
