% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchGenesForSet.R
\name{fetchGenesForSet}
\alias{fetchGenesForSet}
\title{Fetch genes for a single set}
\usage{
fetchGenesForSet(
  species,
  set,
  cache = NULL,
  overwrite = FALSE,
  force.remote = FALSE
)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{set}{Integer identifying the set of interest, as an index into the list of all sets for this species 
(i.e., the return value of \code{\link{fetchGenesForAllSets}}).}

\item{cache}{String containign the path to a cache directory.
If \code{NULL}, a cache location is automatically chosen.}

\item{overwrite}{Logical scalar indicating whether any cached file should be overwritten.}

\item{force.remote}{Logical scalar indicating whether to force a remote query.}
}
\value{
Integer vector containing the identities of the genes in that set, 
where each integer is an index into \code{\link{fetchAllGenes}}.
}
\description{
Fetch genes for a single set in the \pkg{gesel} index.
}
\examples{
first.set <- fetchGenesForSet("9606", 1)

# Genes in the first set:
fetchAllGenes("9606")$symbol[first.set]

}
\author{
Aaron Lun
}
