% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchGenesForSomeSets.R
\name{fetchGenesForSomeSets}
\alias{fetchGenesForSomeSets}
\title{Fetch genes for some sets}
\usage{
fetchGenesForSomeSets(species, sets, config = NULL)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{sets}{Integer vector containing set indices.
Each set index refers to a row in the data frame returned by \code{\link{fetchAllSets}}.}

\item{config}{Configuration list, typically created by \code{\link{newConfig}}.
If \code{NULL}, the default configuration is used.}
}
\value{
List of integer vectors.
Each vector corresponds to a set in \code{sets} and contains the identities of its member genes.
Each gene is defined by its gene index, which refers to a row of the data frame returned by \code{\link{fetchAllGenes}}.
}
\description{
Fetch genes for some sets in the Gesel database.
This can be more efficient than \code{\link{fetchGenesForAllSets}} if only a few sets are of interest.
}
\examples{
first.set <- fetchGenesForSomeSets("9606", 1)
str(first.set)

# Genes in the first set:
gene.symbols <- fetchAllGenes("9606")$symbol
head(gene.symbols[first.set[[1]]])

}
\author{
Aaron Lun
}
