% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchGenesForSomeSets.R
\name{fetchGenesForSomeSets}
\alias{fetchGenesForSomeSets}
\title{Fetch genes for some sets}
\usage{
fetchGenesForSomeSets(
  species,
  sets,
  fetch.file = downloadDatabaseFile,
  fetch.file.args = list(),
  fetch.range = downloadDatabaseRanges,
  fetch.range.args = list(),
  use.preloaded = TRUE
)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{sets}{Integer vector containing set indices.
Each set index refers to a row in the data frame returned by \code{\link{fetchAllSets}}.}

\item{fetch.file}{Function that accepts the name of the file in the Gesel database and returns an absolute path to the file.
If \code{NULL}, it defaults to \code{\link{downloadIndexFile}}.}

\item{fetch.file.args}{Named list of arguments to pass to \code{fetch.file}.}

\item{fetch.range}{Function that accepts at least two arguments - 
the name of the file in the Gesel database, and an integer vector of length 2 containing the zero-indexed half-open byte range to extract from the file
(see \code{\link{downloadIndexRange}} for details).
It should return a string containing the contents of the specified byte range.
If \code{NULL}, it defaults to \code{\link{downloadIndexRange}}.}

\item{fetch.range.args}{Named list of arguments to pass to \code{fetch.file}.}

\item{use.preloaded}{Logical scalar indicating whether to use the preloaded value from a previous call to \code{\link{fetchGenesForAllSets}}.}
}
\value{
List of integer vectors.
Each vector corresponds to a set in \code{sets} and contains the identities of its member genes.
Each gene is defined by its gene index, which refers to a row of the data frame returned by \code{\link{fetchAllGenes}}.
}
\description{
Fetch genes for some sets in the Gesel database.
This can be more efficient than \code{\link{fetchGenesForAllSets}} if only a few sets are of interest.
}
\examples{
first.set <- fetchGenesForSomeSets("9606", 1)
str(first.set)

# Genes in the first set:
gene.symbols <- fetchAllGenes("9606")$symbol
head(gene.symbols[first.set[[1]]])

}
\author{
Aaron Lun
}
