% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSetsForAllGenes.R
\name{fetchSetsForAllGenes}
\alias{fetchSetsForAllGenes}
\title{Fetch sets for all genes}
\usage{
fetchSetsForAllGenes(
  species,
  fetch = downloadDatabaseFile,
  fetch.args = list(),
  use.preloaded = TRUE
)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{fetch}{Function that accepts the name of a Gesel database file and returns an absolute path to that file.}

\item{fetch.args}{Named list of arguments to pass to \code{fetch}.}

\item{use.preloaded}{Logical scalar indicating whether to use the preloaded value from a previous call to this function.}
}
\value{
List of integer vectors.
Each vector corresponds to a gene in the same order as \code{\link{fetchAllGenes}}.
Each vector contains the identities of the sets that contain that gene, 
where each integer is a set index that refers to a row of the data frame returned by \code{\link{fetchAllSets}}.
}
\description{
Fetch the identities of the sets that contain each gene in the Gesel database.
}
\examples{
all.genes <- fetchSetsForAllGenes("9606")
length(all.genes)

# Sets containing the first gene:
head(fetchAllSets("9606")[all.genes[[1]],])

# Details about the first gene:
fetchAllGenes("9606")$symbol[1]

}
\author{
Aaron Lun
}
