% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSomeSets.R
\name{fetchSetSizes}
\alias{fetchSetSizes}
\title{Size of gene sets}
\usage{
fetchSetSizes(species, fetch.file = NULL, fetch.file.args = list())
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{fetch.file}{Function that accepts the name of the file in the Gesel database and returns an absolute path to the file.}

\item{fetch.file.args}{Named list of arguments to pass to \code{fetch.file}.}
}
\value{
Integer vector containing the size of each set (i.e., the number of gene sets).
}
\description{
Quickly get the size of the sets in the Gesel database.
This is more efficient than \code{\link{fetchAllSets}} when only the sizes are of interest.
}
\examples{
head(fetchSetSizes("9606"))

}
\author{
Aaron Lun
}
