% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSomeSets.R
\name{fetchSomeSets}
\alias{fetchSomeSets}
\title{Fetch some sets}
\usage{
fetchSomeSets(
  species,
  sets,
  fetch.file = downloadDatabaseFile,
  fetch.file.args = list(),
  fetch.range = downloadDatabaseRanges,
  fetch.range.args = list()
)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{sets}{Integer vector of set indices, where each set index refers to a row in the data frame returned by \code{\link{fetchAllSets}}.}

\item{fetch.file}{Function that accepts the name of the file in the Gesel database and returns an absolute path to the file.}

\item{fetch.file.args}{Named list of arguments to pass to \code{fetch.file}.}

\item{fetch.range}{Function that accepts at least two arguments - 
the name of the file in the Gesel database, and an integer vector of length 2 containing the zero-indexed half-open byte range to extract from the file
(see \code{\link{downloadDatabaseRanges}} for details).
It should return a string containing the contents of the specified byte range.}

\item{fetch.range.args}{Named list of arguments to pass to \code{fetch.file}.}
}
\value{
Data frame with the same columns as the return value of \code{\link{fetchAllSets}},
where each row corresponds to an entry of \code{sets}.
}
\description{
Fetch the details of some gene sets from the Gesel database.
This can be more efficient than calling \code{\link{fetchAllSets}} when only a few sets are of interest.
}
\examples{
fetchSomeSets("9606", 1)

}
\author{
Aaron Lun
}
