% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlappingSets.R
\name{findOverlappingSets}
\alias{findOverlappingSets}
\title{Find sets overlapping a list of genes}
\usage{
findOverlappingSets(species, genes, num.workers = 4L, more.args = list())
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{genes}{Integer vector representing a list of genes.
Each entry is an index into one of the lists returned by \code{\link{fetchAllGenes}}.}

\item{num.workers}{Integer scalar specifying the number of workers to use for parallelizing calls to \code{\link{fetchSetsForGene}}.}

\item{more.args}{Further arguments to pass to \code{\link{fetchSetsPerGene}}.}
}
\value{
A list containing:
\itemize{
\item \code{overlap}, a data frame of the overlapping sets.
Each row represents a set and has the columns \code{set} and \code{count}.
The former contains the identity of the set as an index into \code{\link{fetchAllSets}}.
The latter contains the number of overlaps with \code{genes} for that set.
\item \code{present}, an integer scalar containing the number of genes in \code{genes} that are present in at least one set in the Gesel index for \code{species}.
This can be used to remove genes outside of the universe prior to testing for enrichment.
}
}
\description{
Find all sets overlapping any gene in a user-supplied list, and return the number of overlaps per set.
}
\examples{
overlaps <- findOverlappingSets("9606", 1:10)
head(overlaps$overlap)

}
\author{
Aaron Lun
}
