% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlappingSets.R
\name{findOverlappingSets}
\alias{findOverlappingSets}
\title{Find sets overlapping a list of genes}
\usage{
findOverlappingSets(species, genes, more.args = list())
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{genes}{Integer vector containing gene indices.
Each gene index refers to a row of the data frame returned by \code{\link{fetchAllGenes}}.}

\item{more.args}{Further arguments to pass to \code{\link{fetchSetsForSomeGenes}}.}
}
\value{
A list containing:
\itemize{
\item \code{overlap}, a data frame of the overlapping sets.
Each row represents a set and has the integer columns \code{set} and \code{count}.
The former is a set index that refers to a row of the data frame returned by \code{\link{fetchAllSets}}.
The latter specifies the number of overlaps with \code{genes} for that set.
\item \code{present}, an integer scalar containing the number of genes in \code{genes} that are present in at least one set in the Gesel database for \code{species}.
This can be used to remove genes outside of the universe prior to testing for enrichment.
}
}
\description{
Find all sets overlapping any gene in a user-supplied list, and return the number of overlaps per set.
}
\examples{
overlaps <- findOverlappingSets("9606", 1:10)
head(overlaps$overlap)

}
\author{
Aaron Lun
}
