% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlappingSets.R
\name{findOverlappingSets}
\alias{findOverlappingSets}
\title{Find sets overlapping a list of genes}
\usage{
findOverlappingSets(species, genes, counts.only = TRUE, config = NULL)
}
\arguments{
\item{species}{String containing the NCBI taxonomy ID of the species of interest.}

\item{genes}{Integer vector containing gene indices.
Each gene index refers to a row of the data frame returned by \code{\link{fetchAllGenes}}.}

\item{counts.only}{Logical scalar indicating whether to only report the number of overlapping genes for each set.}

\item{config}{Configuration list, typically created by \code{\link{newConfig}}.
If \code{NULL}, the default configuration is used.}
}
\value{
A list containing:
\itemize{
\item \code{overlap}, a data frame of the overlapping sets.
Each row represents a set that is identified by the set index in the \code{set} column.
(This set index refers to a row of the data frame returned by \code{\link{fetchAllSets}}.)
It also has:
\itemize{
\item The \code{count} column, if \code{counts.only=TRUE}.
This specifies the number of overlaps between the genes in the set and those in \code{genes}.
\item The \code{genes} column, if \code{counts.only=FALSE}.
This is a list that contains the entries of \code{genes} that overlap with those in the set.
}
\item \code{present}, an integer scalar containing the number of genes in \code{genes} that are present in at least one set in the Gesel database for \code{species}.
}
}
\description{
Find all sets overlapping any gene in a user-supplied list, and return the number of overlaps per set.
}
\details{
The \code{present} number should be used as the number of draws when performing a hypergeomtric test for gene set enrichment
(see \code{\link{phyper}}), instead of \code{length(genes)}.
It ensures that genes outside of the Gesel universe are ignored, e.g., due to user error, different genome versions.
Otherwise, unknown genes would inappropriately increase the number of draws and inflate the enrichment p-value.
}
\examples{
overlaps <- findOverlappingSets("9606", 1:10)
head(overlaps$overlap)

}
\author{
Aaron Lun
}
