% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchGenes.R
\name{searchGenes}
\alias{searchGenes}
\title{Search for genes}
\usage{
searchGenes(species, genes, types = NULL, ignore.case = TRUE, config = NULL)
}
\arguments{
\item{species}{String specifying the taxonomy ID of the species of interest.}

\item{genes}{Character vector of gene names of any type specified in \code{types}.}

\item{types}{Character vector specifying the types of gene names to return.
This is typically one or more of \code{"symbol"}, \code{"entrez"}, and \code{"ensembl"},
defaulting to all of them.}

\item{ignore.case}{Logical scalar indicating whether case should be ignored.}

\item{config}{Configuration list, typically created by \code{\link{newConfig}}.
If \code{NULL}, the default configuration is used.}
}
\value{
List of length equal to \code{genes}.
Each entry is an integer vector of gene indices that refer to rows of the data frame returned by \code{\link{fetchAllGenes}};
these rows represent the genes that match to the corresponding entry of \code{genes}.
}
\description{
Search for genes by converting gene identifiers to gene indices.
}
\examples{
searchGenes("9606", c("SNAP25", "NEUROD6", "ENSG00000139618"))

}
\author{
Aaron Lun
}
